seperate server logic out
